requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.10.0
openpyxl>=3.0.0
fastapi>=0.104.0
uvicorn>=0.23.0
requests>=2.31.0
flask>=2.3.0
werkzeug>=2.3.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
pydantic>=2.0.0
pyyaml>=6.0
pytest-benchmark
pytest
httpx
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv